Implemented Transaction Fragment, Tabs (History and Upcomming), Transaction Adapters, HistoryUpcomingFragment, item_transaction

+++ app/src/main/java/com/example/jdocter/lendme/DetailPostActivity.java
@@ -89,7 +89,7 @@ public class DetailPostActivity extends AppCompatActivity {
                     }
                     tvDescription.setText(post.getDescription());
                     tvTitleItem.setText(post.getItem());
-                    tvPrice.setText("$"+Integer.toString(post.getPrice()));
+                    tvPrice.setText("$"+Float.toString(post.getPrice()));
 
                 } else {
                     e.printStackTrace();
